backend/requirements.txt
pydantic[dotenv]
numpy
pinecone-client
typing-extensions
python-dotenv
loguru
pytest
pytest-asyncio
numpy>=1.20
pydantic>=2.5.0
python-dotenv>=0.20
click>=8.0
loguru>=0.5.0 ; sys_platform == 'win32'
colorlog>=6.8.2 ; sys_platform != 'win32'
scipy>=1.8.0
pytest>=7.0
openai>=1.12.0
fastapi
uvicorn
